Add files via upload
9:33 PM
12/3/2020

Changed a lot of the forms to give visual feedback when the button is hovered over. 90% of them are done.

I have figured out how to serialize a player, and am able to reproduce the tournament roster from a text file.
Am currently planning out how I would recursively serialize games in a match, and the logic is a bit tricky for me.
Just backing up because a lot of progress has been made.