Add files via upload
12/12/2020 12:43 AM

Will write out a more detailed update later, but basically redid the serialization system using QFileInfo, QFile, and QDir to make it MUCH cleaner, more intuitive, and reusable for other tournament formats. 

Changed the Tournament Roster from a listWidget to a tableWidget to display a player's name AND points (this is to display tournament standings whenever needed).

Those are the 2 biggest changes implemented.