Action : Modified 'MainMenu.cpp' so that on 'ExitButton' clicked it closes the program like it should.
--------------------------------------------------------------------------------

Action : Modified SelectRoundDialog to become much less convoluted. Before, the flow of work would be

1. Tournament Menu : Choose Round
2. Tournament Menu creates SelectRoundDialog().
3. User chooses Round.
4. AddToRoundDialog is created from WITHIN SelectRoundDialog.
5. User modifies players in round if necessary.
6. Both forms close and return to TournamentMenu.

This was unneccesary and made it harder to track for errors.

Since the true purpose of SelectRoundDialog is just to return the index of the chosen round, "int getSelectedRound()" was added to SelectRoundDialog class.
This eliminates the need to open another dialog within this dialog. Now, the chosen index is simply returned to TournamentMenu via getSelectedRound(). Now all necessary steps
to select a round from a tournament are called from a TournamentMenu object.

Now the flow of work is :
1. Tournament Menu : Choose Round
2. Tournament Menu creates SelectRoundDialog.
3. User chooses Round.
4. Index of round is returned to TournamentMenu.
5. TournamentMenu creates AddToRoundDialog.
6. User modifies players in round if necessary.
7. Form closes. TournamentMenu closes, RoundMenu is created.

These changes make the logic flow simpler and easier to debug if needed.

--------------------------------------------------------------------------------

Action : Changed names of listWidgets in SelectRoundDialog to clarify what data they display.

--------------------------------------------------------------------------------

Action : Removed printing of rounds from the SelectRoundDialog constructor and placed them into a function. The constructor now calls the function instead.
Action : Created a function that displays the players in a round which is called whenever a round from RoundsWidget is clicked.