Changes Made:
* Added parameter, "playerRoster" to each instance of deserialize"X"Players() (where X is Tournament, Round, Match, or Game).

* Edited the functions so that when a player object is created, it is cross referenced with the roster of players, and if that player already exists, the newly
created object's shared pointer is reset and assigned to the already existing object.

Reason:

During deserialization when creating players, all std::shared_ptr<Player> newPlayers created were not made to point to any previously existing copies of that object.
This effectively makes all shared_ptrs different objects that happen to have the same name.

Example:

"John" is an object, with shared_ptrs to "John" in TournamentRoster, RoundRoster, MatchRoster, and GameRoster.
When each "John" shared_ptr is created during deserialization, it is not made to point to any preexisting "John" objects.
Resulting in 4 different "John" objects with the same name.

Result:

All references to a player in any part of the program refer to the same object.
(This hereby makes it problematic for player objects to have the same name, will need to implement a safeguard to assure no duplicate names are allowed.)