Modified code to no longer use scoreSort();
This caused problems as the gameResults map in a game uses the names of the pointers in the playersInGames vector as keys through a for loop.
The scoreSort(); function reassignes the values of all pointers in order to sort them in ascending order in an array.
This causes problems when loading tournaments as the names inside a playersInGames vector are changed, and the already loaded resultsMap will try to use the name of a player
potentially not in the game as a search key, which causes the code to break.

Also modified SelectRoundDialog to display list of matches in a round instead of players in a round. (More useful).