Change: In Tournament, Round, Match, and Game classes, the serialization methods were changed to remove all instances of file.remove(); Reason: QT's File class has a destructor that is called when the File goes out of scope (I did not know this). The file.remove() statements are redudant, and would therefore break serialization causing tournament data to corrupt on save attempt.
Change: Modified MatchMenu style sheets. deleteGameButton and finishGameButton are now both disabled and set to gray at intialization, and
are only enabled/colored in when a valid game is clicked from the widget.
Reason: It gives the user more visual cues/makes menu more intuitive.

Change: Modified RoundMenu style sheet, selectMatchButton is now disabled and set to gray at intialization, and
are only enabled/colored in when a valid match is clicked from the widget.
Reason: It gives the user more visual cues/makes menu more intuitive.

To-do list:
privatize public member variables that don't need to be public
(this was done early on to ease development. it is unprofessional but i was too excited.)

potentially rename member variables to be clearer.

include installation binaries.